Allow team names of up to 255 characters (#670)
This PR changes the maximum allowed length of a Gitea team from 30 to 255 characters. This reflects the team names allowed in Gitea as pr. [1.22.2](https://github.com/go-gitea/gitea/releases/tag/v1.22.2). The change is described as _"Allow org team names of length 255 in create team form"_.

I would prefer it, if the _SDK_. did not check the length of team names as I believe the responsibility of rejecting invalid team creation requests should be done by Gitea itself.

Co-authored-by: tobias.petersen <tobias.petersen@unity3d.com>
Reviewed-on: https://gitea.com/gitea/go-sdk/pulls/670
Reviewed-by: delvh <dev.lh@web.de>
Reviewed-by: techknowlogick <techknowlogick@noreply.gitea.com>
Co-authored-by: tobiasbp <tobiasbp@noreply.gitea.com>
Co-committed-by: tobiasbp <tobiasbp@noreply.gitea.com>