(feat) support more query params for /repos/{owner}/{repo}/commits (#668)
This PR adds the ability to set some of the missing (but supported) options to the `ListRepoCommits` function which are performance critical. Without the ability to set these flags, the `ListRepoCommits` takes a really long time to just return a list of commits.

I referenced the following documentation for this patch <https://docs.gitea.com/api/1.22/#tag/repository/operation/repoGetAllCommits>.

Co-authored-by: techknowlogick <techknowlogick@noreply.gitea.com>
Reviewed-on: https://gitea.com/gitea/go-sdk/pulls/668
Reviewed-by: techknowlogick <techknowlogick@noreply.gitea.com>
Co-authored-by: Leon Jacobs <leonja511@gmail.com>
Co-committed-by: Leon Jacobs <leonja511@gmail.com>