Expand Repository Push Mirrors (#722)
This PR expands the existing support for managing [Repository Push Mirrors](https://docs.gitea.com/usage/repo-mirror#pushing-to-a-remote-repository). It includes the following additions:

- **ListPushMirrors** (`GET /repos/{owner}/{repo}/push_mirrors`): gets all push mirrors of a repository
- **GetPushMirrorByRemoteName** (`GET /repos/{owner}/{repo}/push_mirrors/{name}`): get a push mirror of the repository by remote name
- **DeletePushMirror** (`DELETE /repos/{owner}/{repo}/push_mirrors/{name}`): deletes a push mirror from a repository by remote name

A new test file has been added to ensure the functionality works as expected. The tests cover the creation, listing, retrieval, and deletion of push mirrors, validating the full lifecycle of the feature.

Reviewed-on: https://gitea.com/gitea/go-sdk/pulls/722
Reviewed-by: Lunny Xiao <xiaolunwen@gmail.com>
Co-authored-by: Madh93 <mhdez@protonmail.com>
Co-committed-by: Madh93 <mhdez@protonmail.com>